printk: add real time printing option
To align kernel and android log, we replace cpu time with real time. The
function is not enabled by default, userspace should explicit write 1 to
/sys/module/printk/parameters/walltime to enable this function.

Some people may doubt that time_to_tm() have some performance fallback,
but in end user environment, the log printing is disabled by default. So
it's harmless to enable this function.

Signed-off-by: WEN Pingbo <pingbo.wen@linaro.org>